+++ linx/utils/types/productList.ts
