+++ linx/utils/types/shared-types/labelOptions.ts
@@ -1,15 +1,15 @@
 export interface Tag {
-    Label: string;
-    TagID: number;
-  }
-  
-  export interface SortOption {
-    Alias: string;
-    Label: string;
-    Selected: boolean;
-  }
-  
-  export interface Spell {
-    Collation: string;
-    Options: unknown[];
-  }
+  Label: string;
+  TagID: number;
+}
+
+export interface SortOption {
+  Alias: string;
+  Label: string;
+  Selected: boolean;
+}
+
+export interface Spell {
+  Collation: string;
+  Options: unknown[];
+}
