+++ linx/utils/types/shared-types/productDefinition.ts
@@ -1,0 +1,9 @@
+export interface ProductDefinition {
+    IsJoinable: boolean;
+    JoinDisplayText: null;
+    JoinProductBehavior: number;
+    JoinPropertyMetadataID: null;
+    JoinStockKeepUnitPatternLength: null;
+    ProductDefinitionID: number;
+  }
+
